# docs/industrial/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_knowledge_bases__kb_id__documents_post": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_knowledge_bases__kb_id__documents_post",
        "type": "object"
      },
...
      "DocumentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "DocumentPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "JobPage",
        "type": "object"
      },
...
      "LegacyView": {
        "properties": {
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          }
        },
        "required": [
          "revision_id"
        ],
        "title": "LegacyView",
        "type": "object"
      },
...
      "MemberView": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "display_name",
          "role"
        ],
        "title": "MemberView",
        "type": "object"
      },
...
      "ReparseView": {
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "revision_id",
          "job_id",
          "state"
        ],
        "title": "ReparseView",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-tracedesk-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "TraceDesk Team API",
    "version": "0.3.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/document-revisions/{revision_id}/file": {
      "get": {
        "operationId": "original_file_api_v1_document_revisions__revision_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Original File"
      }
    },
    "/api/v1/document-revisions/{revision_id}/reparse": {
      "post": {
        "operationId": "reparse_api_v1_document_revisions__revision_id__reparse_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Reparse"
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "retry_wait",
                    "succeeded",
                    "failed",
                    "cancelled",
                    "superseded"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Jobs"
      }
    },
...
    "/api/v1/knowledge-bases/{kb_id}/documents": {
      "get": {
        "operationId": "list_documents_api_v1_knowledge_bases__kb_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Documents"
      },
      "post": {
        "operationId": "upload_api_v1_knowledge_bases__kb_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_knowledge_bases__kb_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Upload"
      }
    },
...
    "/api/v1/knowledge-bases/{kb_id}/members/{user_id}": {
      "delete": {
        "operationId": "remove_kb_member_api_v1_knowledge_bases__kb_id__members__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Remove Kb Member"
      },
      "get": {
        "operationId": "member_api_v1_knowledge_bases__kb_id__members__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Member"
      },
      "put": {
        "operationId": "kb_member_api_v1_knowledge_bases__kb_id__members__user_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Kb Member"
      }
    },
    "/api/v1/legacy/sources/{legacy_doc_id}": {
      "get": {
        "operationId": "legacy_source_api_v1_legacy_sources__legacy_doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "legacy_doc_id",
            "required": true,
            "schema": {
              "title": "Legacy Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Legacy Source"
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/knowledge-bases": {
      "get": {
        "operationId": "knowledge_bases_api_v1_workspaces__workspace_id__knowledge_bases_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KBPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Knowledge Bases"
      },
      "post": {
        "operationId": "create_kb_api_v1_workspaces__workspace_id__knowledge_bases_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKBBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create Kb"
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/users": {
      "get": {
        "operationId": "users_api_v1_workspaces__workspace_id__users_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Users"
      }
    },
...
  }
}